Split up layer tests so that we have 1:1 between modules and test modules.
